[Fix] Constraints library was not efficient enough to work with up to 8 players. Switched this out for scipy
