+++ .github/workflows/devenv-checks.yml
@@ -16,7 +16,7 @@ jobs:
       - uses: cachix/cachix-action@v14
         with:
           name: devenv
-      - name: Install devenv.sh
+      - name: Install devenv
         run: nix profile install nixpkgs#devenv
       - name: Run pre-commit checks
         run: devenv test
