+++ backend/games_backend/games/quantum/constants.py
@@ -1,0 +1,5 @@
+# In Quantum Go Fish, the number of suits always equals the number of players
+# Each suit is created by the first player to ask for it
+UNKNOWN_SUIT = -1
+CARDS_PER_SUIT = 4
+DEFAULT_HAND_SIZE = 4
